refactor(core): adopt_shared, keep.rs fold in (KEN-549) [no-changelog]
adopt.rs and adopt_shared.rs imported each other. Pointing the imports
one way means moving read_tree and local_item_path into adopt_shared,
which would then hold the local-path rule and the capture walker, neither
about shared folders, and adopt's own capture_ops would import its walker
from the "shared" file. That is the ping-pong, so the shared-folder branch
goes back into adopt.rs beside the plain capture it mirrors. Its tests
move to adopt/tests.rs. exits.rs and tree_plan.rs now name
adopt::shared_tools and adopt::link_target.

detach/keep.rs said in its own header that the declaration flip lives in
the parent; remove() and source() are the two arms of one unsubscribe
with the same edited-refusal preamble, so keep.rs folds into detach.rs.

desired/rebuild.rs stays: it has its own types and tests and reads
desired_state as a black box, so its reader never needs desired.rs open.
Its header no longer narrates the split.

Bodies move byte-for-byte; the diff is mod, use, visibility and
doc-comment lines plus one rustfmt collapse of local_item_path after its
visibility shrank. No behavior change. Items that were pub(super) or
pub(crate) for a sibling now in the same file are private; shared_tools
and link_target stay pub(super) for exits.rs and tree_plan.rs.

adopt.rs (539 -> 726) and detach.rs (new row, 615) rise under
RATCHET_RAISE=1 (size-ratchet: merging fragments). The emptied files are
deleted in this diff and had no rows.
